+++ app/src/main/java/co/kr/woowahan_accountbook/data/datasource/local/classification/ClassificationDataSourceImpl.kt
@@ -1,4 +1,4 @@
-package co.kr.woowahan_accountbook.data.datasource.local.payment
+package co.kr.woowahan_accountbook.data.datasource.local.classification
 
 import android.content.ContentValues
 import android.database.sqlite.SQLiteDatabase
